refactor: extract share icon and text logic into useShareDraw hook
- Renamed `src/hooks/useShareDraw.ts` to `.tsx` to support JSX.
- Extracted `shareIcon` generating function and `shareText` logic into `useShareDraw`.
- Refactored `DrawSection`, `DrawActions`, and `SavedActions` to use the returned `shareIcon` and `shareText` from `useShareDraw`, reducing code duplication.
- Applied consistent disabled logic across the dropdowns and button groups.
